move demo model class to common lib (#182)
* move demo model class to common lib

* change text-based to text_based

* resolve comment

* add multinode strageties
