+++ demo/__init__.py
@@ -14,4 +14,14 @@
  limitations under the License.
  """
 
-from . import lightning
+from demo.lightning.checkpoint.multinode.train import (DemoTransformer,
+                                                       init_processes)
+from demo.lightning.checkpoint.multinode.strategies import (
+    DatafluxFSDPStrategy, FSSpecFSDPStrategy, LoadFromBootDiskFSDP)
+from demo.lightning.text_based.demo_model import (TextDemoModel, format_data)
+
+__all__ = [
+    "DemoTransformer", "init_processes", "DatafluxFSDPStrategy",
+    "FSSpecFSDPStrategy", "LoadFromBootDiskFSDP", "TextDemoModel",
+    "format_data"
+]
