+++ demo/lightning/checkpoint/__init__.py
@@ -14,4 +14,11 @@
  limitations under the License.
  """
 
-from . import multinode
+from .multinode.train import (DemoTransformer, init_processes)
+from .multinode.strategies import (DatafluxFSDPStrategy, FSSpecFSDPStrategy,
+                                   LoadFromBootDiskFSDP)
+
+__all__ = [
+    "DemoTransformer", "init_processes", "DatafluxFSDPStrategy",
+    "FSSpecFSDPStrategy", "LoadFromBootDiskFSDP"
+]
