+++ demo/lightning/checkpoint/multinode/__init__.py
@@ -14,4 +14,11 @@
  limitations under the License.
  """
 
-from .train import *
+from .train import (DemoTransformer, init_processes)
+from .strategies import (DatafluxFSDPStrategy, FSSpecFSDPStrategy,
+                         LoadFromBootDiskFSDP)
+
+__all__ = [
+    "DemoTransformer", "init_processes", "DatafluxFSDPStrategy",
+    "FSSpecFSDPStrategy", "LoadFromBootDiskFSDP"
+]
