+++ demo/lightning/text_based/__init__.py
@@ -1,0 +1,3 @@
+from .demo_model import (TextDemoModel, format_data)
+
+__all__ = ["TextDemoModel", "format_data"]
