+++ kokoro/bench.sh
@@ -48,7 +48,7 @@ function install_requirements() {
     pip install -r requirements.txt
 
     echo Installing parquet demo requirements.
-    pip install -r ./demo/lightning/text-based/distributed/requirements.txt
+    pip install -r ./demo/lightning/text_based/distributed/requirements.txt
 
     echo Installing image training demo requirements.
     pip install -r ./demo/lightning/image-segmentation/requirements.txt
@@ -60,7 +60,7 @@ function install_requirements() {
 function run_benchmarks(){
     echo Running benchmarks...
     echo Running parquet-text benchmark.
-    python3 -u ./demo/lightning/text-based/distributed/model.py --local --project=dataflux-project --bucket=fineweb-df-benchmark --num-workers=2 --num-nodes=1 --devices=5 --batch-size=512 --epochs=5 --limit-train-batches=1000 --log-level=ERROR;
+    python3 -u ./demo/lightning/text_based/distributed/model.py --local --project=dataflux-project --bucket=fineweb-df-benchmark --num-workers=2 --num-nodes=1 --devices=5 --batch-size=512 --epochs=5 --limit-train-batches=1000 --log-level=ERROR;
     echo Running image-segmentation benchmark.
     python3 -u ./demo/lightning/image-segmentation/train.py --local --benchmark --gcp_project=dataflux-project --gcs_bucket=dataflux-demo-public --images_prefix=image-segmentation-dataset/images --labels_prefix=image-segmentation-dataset/labels --num_nodes=1 --num_devices=5 --epochs=2;
     echo Installing checkpointing demo requirements.
