+++ setup.py
@@ -12,7 +12,8 @@ setup(
     name="dataflux-pytorch",
     packages=[
         "dataflux_core", "dataflux_pytorch", "dataflux_pytorch.lightning",
-        "dataflux_pytorch.multipart_upload"
+        "dataflux_pytorch.multipart_upload", "demo.lightning.text_based",
+        "demo.lightning.checkpoint.multinode"
     ],
     package_dir={"dataflux_core": "dataflux_client_python/dataflux_core"},
     install_requires=dependencies,
